Move functions to test modules to _utils.py
