+++ skrub/tests/test_datetime_encoder.py
@@ -7,7 +7,6 @@ import pytest
 from numpy.testing import assert_allclose, assert_array_equal
 from pandas.testing import assert_frame_equal
 
-from skrub._dataframe._namespace import is_namespace_polars
 from skrub._dataframe._polars import POLARS_SETUP
 from skrub._datetime_encoder import (
     TIME_LEVELS,
@@ -15,6 +14,7 @@ from skrub._datetime_encoder import (
     _is_pandas_format_mixed_available,
     to_datetime,
 )
+from skrub._utils import is_namespace_polars
 
 MODULES = [pd]
 ASSERT_TUPLES = [(pd, assert_frame_equal)]
