+++ skrub/tests/test_joiner.py
@@ -4,8 +4,8 @@ from numpy.testing import assert_array_equal
 from pandas.testing import assert_frame_equal
 
 from skrub import Joiner
-from skrub._dataframe._namespace import is_namespace_polars
 from skrub._dataframe._polars import POLARS_SETUP
+from skrub._utils import is_namespace_polars
 
 MODULES = [pd]
 ASSERT_TUPLES = [(pd, assert_frame_equal)]
